fixed type for flags

+++ code/core/log_reader.c
@@ -7,13 +7,13 @@
 int log_reader_init(void) {
     int fd;
 
-    uint64_t init_flags =
+    unsigned int init_flags =
         FAN_CLASS_NOTIF |             // notifications uniquement (pas de blocage)
         FAN_NONBLOCK |                // mode non-bloquant
         FAN_UNLIMITED_QUEUE |         // pas de limite de taille de file d'attente
         FAN_UNLIMITED_MARKS;          // pas de limite de marques
 
-    uint64_t init_flags = O_RDONLY | O_LARGEFILE; 
+    unsigned int init_flags = O_RDONLY | O_LARGEFILE; 
 
     fd = fanotify_init(init_flags, event_flags);
